+++ packages/thread-br/src/credentials/index.ts
@@ -1,8 +1,0 @@
-/**
- * @license
- * Copyright 2025-2026 Open Home Foundation
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export type { ThreadNetworkCredentials } from "./ThreadNetworkCredentials.js";
-export { ThreadCredentialsRegistry } from "./ThreadCredentialsRegistry.js";