+++ packages/thread-br/src/crypto/index.ts
@@ -1,9 +1,0 @@
-/**
- * @license
- * Copyright 2025-2026 Open Home Foundation
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export { AesCmac, AesCmacPrf128 } from "./AesCmacPrf128.js";
-export { pbkdf2AesCmac } from "./Pbkdf2AesCmac.js";
-export { Pskc } from "./Pskc.js";