+++ packages/thread-br/src/dataset/index.ts
@@ -1,9 +1,0 @@
-/**
- * @license
- * Copyright 2025-2026 Open Home Foundation
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export { OperationalDataset } from "./OperationalDataset.js";
-export { SecurityPolicy } from "./SecurityPolicy.js";
-export { MeshCopTlvType, MeshCopTlvTypeName } from "./meshcopTlvTypes.js";